+++ src/libprofit/VERSION
@@ -1,1 +1,1 @@
-0.7
+0.7.1
